+++ packages/generator/src/action-function-import/action.ts
@@ -29,7 +29,7 @@ export function actionImportFunction(
       [
         `${actionImport.description}\n`,
         '@param parameters - Object containing all parameters for the action import.',
-        '@returns A request builder that allows to overwrite some of the values and execute the resultng request.'
+        '@returns A request builder that allows to overwrite some of the values and execute the resulting request.'
       ].join('\n')
     ]
   };