+++ test-packages/test-services-e2e/TripPin/microsoft-o-data-service-sample-trippin-in-memory-models-service/action-imports.js
@@ -11,7 +11,7 @@ var core_1 = require("@sap-cloud-sdk/core");
  * Reset Data Source.
  *
  * @param parameters - Object containing all parameters for the action import.
- * @returns A request builder that allows to overwrite some of the values and execute the resultng request.
+ * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
  */
 function resetDataSource(parameters) {
     var params = {};