+++ test-packages/test-services-e2e/v4/admin-service/index.ts
@@ -7,4 +7,5 @@ export * from './TestEntity';
 export * from './TestEntityLink';
 export * from './TestEntityRequestBuilder';
 export * from './TestEntityLinkRequestBuilder';
+export * from './function-imports';
 export * from './BatchRequest';